move the compute norm function to the compute norm script
